last touches to messages
