+++ src/app/messages/layout.tsx
@@ -1,37 +1,28 @@
 "use client";
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { useState, useEffect } from "react";
+import { useState } from "react";
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      staleTime: 30000, // 30 seconds
+      refetchOnWindowFocus: false,
+      retry: false,
+    },
+  },
+});
 
 export default function MessagesLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const [mounted, setMounted] = useState(false);
-  const [queryClient] = useState(
-    () =>
-      new QueryClient({
-        defaultOptions: {
-          queries: {
-            staleTime: 30 * 1000, // Consider data stale after 30 seconds
-            refetchInterval: false, // Disable automatic refetching
-            refetchOnWindowFocus: false, // Disable refetch on window focus
-            retry: false, // Disable retries
-          },
-        },
-      })
-  );
-
-  useEffect(() => {
-    setMounted(true);
-  }, []);
-
-  if (!mounted) {
-    return null;
-  }
-
   return (
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    <QueryClientProvider client={queryClient}>
+      <div className="bg-gradient-to-b from-[#F7F1EE] to-[#E5DEDB] min-h-screen">
+        <div className="relative ">{children}</div>
+      </div>
+    </QueryClientProvider>
   );
 }
